Re-implement background content and correct stacking order.

+++ src/core/Content.cpp
@@ -65,11 +65,11 @@ void Content::setDimensions( const QSize& dimensions )
     emit( modified( ));
 }
 
-float Content::getAspectRatio() const
+qreal Content::getAspectRatio() const
 {
-    if ( size_.height() == 0 )
+    if( size_.height() == 0 )
         return 0.f;
-    return (float)size_.width() / (float)size_.height();
+    return (qreal)size_.width() / (qreal)size_.height();
 }
 
 ContentActionsModel* Content::getActions()
